requirements.txt
django
djangorestframework
django-crispy-forms
Pillow
